requirements.txt
asgiref==3.2.10
Django==3.1
django-smart-selects==1.5.9
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
